standardization controller api

+++ src/main/java/com/findhub/finhubbackend/controller/ApiController.java
@@ -3,6 +3,8 @@ package com.findhub.finhubbackend.controller;
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -15,6 +17,12 @@ import com.findhub.finhubbackend.util.Config.ApiPath;
 import com.findhub.finhubbackend.util.Config.Var;
 
 public class ApiController<E, T extends Service<E, S>, S> {
+
+    private E x;
+    private String classSimpleName = x.getClass().getSimpleName();
+    // private String className = x.getClass().getName();
+    // private Class eService = Class.forName(className + "Service");
+    // private Class eStatus = Class.forName(className + "Status");
 
     @Autowired
     protected T service;
@@ -30,23 +38,63 @@ public class ApiController<E, T extends Service<E, S>, S> {
     }
 
     @PostMapping(ApiPath.ADD)
-    public E add(@RequestBody E entity) {
-        return service.add(entity);
+    public ResponseEntity<String> add(@RequestBody E entity) {
+        if (entity == null)
+            return new ResponseEntity<>(
+                    "Failed to add " + classSimpleName, HttpStatus.BAD_REQUEST);
+
+        service.save(entity);
+        return new ResponseEntity<>(
+                "Added new " + classSimpleName + " successfully",
+                HttpStatus.OK);
     }
 
     @PutMapping(ApiPath.UPDATE)
-    public E update(@PathVariable(Var.ID) int id, @RequestBody E entity) {
-        return service.update(id, entity);
+    public ResponseEntity<String> update(@PathVariable(Var.ID) int id, @RequestBody E entity) {
+        if (service.findById(id) == null)
+            return new ResponseEntity<>(
+                    "Failed to update " + classSimpleName + ": " + classSimpleName + "[" + id + "] not found",
+                    HttpStatus.NOT_FOUND);
+
+        if (entity == null)
+            return new ResponseEntity<>(
+                    "Failed to update " + classSimpleName + ": update content is NULL",
+                    HttpStatus.BAD_REQUEST);
+
+        service.update(id, entity);
+        return new ResponseEntity<>(
+                "Updated " + classSimpleName + "[" + id + "] successfully",
+                HttpStatus.OK);
     }
 
     @DeleteMapping(ApiPath.DELETE)
-    public boolean delete(@RequestBody int id) {
+    public ResponseEntity<String> delete(@RequestBody int id) {
         E entity = service.findById(id);
-        return service.delete(entity);
+
+        if (entity == null)
+            return new ResponseEntity<>(
+                    "Failed to delete " + classSimpleName + ": " + classSimpleName + "[" + id + "] not found",
+                    HttpStatus.NOT_FOUND);
+
+        service.delete(entity);
+        return new ResponseEntity<>(
+                "Deleted " + classSimpleName + "[" + id + "] successfully",
+                HttpStatus.OK);
     }
 
     @PostMapping(ApiPath.CHANGE_STATUS)
-    public boolean updateStatus(@RequestBody int id, @RequestBody int status) {
-        return service.updateStatus(id, status);
+    public ResponseEntity<String> updateStatus(@RequestBody int id, @RequestBody int status) {
+        E entity = service.findById(id);
+
+        if (entity == null)
+            return new ResponseEntity<>(
+                    "Failed to update Status[" + status + "] of" + classSimpleName
+                            + ": " + classSimpleName + "[" + id + "] not found",
+                    HttpStatus.NOT_FOUND);
+
+        service.updateStatus(id, status);
+        return new ResponseEntity<>(
+                "Updated Status[" + status + "] of" + classSimpleName + "[" + id + "] successfully",
+                HttpStatus.OK);
     }
 }
